app/requirements.txt
fastapi
uvicorn
celery
redis
ollama
requests
pydantic
qdrant_client==1.9.1
pydantic-settings
mysql-connector-python